Character improvements

+++ amethyst-common/src/main/java/io/github/alyphen/amethyst/common/sprite/Sprite.java
@@ -86,7 +86,7 @@ public class Sprite {
         File metadataFile = new File(directory, "sprite.json");
         Map<String, Object> metadata = FileUtils.loadMetadata(metadataFile);
         File imageFile = new File(directory, "sprite.png");
-        return fromImage(ImageIO.read(imageFile), (int) metadata.get("frame_amount"), (int) metadata.get("frame_delay"), (int) metadata.get("width"), (int) metadata.get("height"));
+        return fromImage(ImageIO.read(imageFile), (int) ((double) metadata.get("frame_amount")), (int) ((double) metadata.get("frame_delay")), (int) ((double) metadata.get("width")), (int) ((double) metadata.get("height")));
     }
 
     public BufferedImage toImage(Sprite sprite) {
